sorting stuff

+++ oop.ueh/OperatorOverloading/OperatorOverloading/Vector3D.cs
@@ -8,9 +8,11 @@ namespace OperatorOverloading
 {
     internal class Vector3D : IVector
     {
-        public int x { get; set; }
-        public int y { get; set; }
-        public int z { get; set; }
+        public double x { get; set; }
+        public double y { get; set; }
+        public double z { get; set; }
+
+        
 
         public Vector3D(int x, int y, int z)
         {
